refactor to highlight custom layer
