+++ custom-layer/base_mnist_model.js
@@ -1,0 +1,50 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+import * as tf from '@tensorflow/tfjs';
+
+/**
+ * Creates and returns a Keras MNIST convolutional model.
+ * See @tensorflow/tfjs-examples/mnist for more.
+ */
+export function baseMnistModel() {
+  // Set up base model
+  const baseModel = tf.sequential();
+  baseModel.add(tf.layers.conv2d({
+    inputShape: [28, 28, 1],
+    kernelSize: 5,
+    filters: 8,
+    strides: 1,
+    activation: 'relu',
+    kernelInitializer: 'varianceScaling'
+  }));
+  baseModel.add(tf.layers.maxPooling2d({poolSize: [2, 2], strides: [2, 2]}));
+  baseModel.add(tf.layers.conv2d({
+    kernelSize: 5,
+    filters: 16,
+    strides: 1,
+    activation: 'relu',
+    kernelInitializer: 'varianceScaling'
+  }));
+  baseModel.add(tf.layers.maxPooling2d({poolSize: [2, 2], strides: [2, 2]}));
+  baseModel.add(tf.layers.flatten());
+  baseModel.add(tf.layers.dense({
+    units: 10,
+    kernelInitializer: 'varianceScaling',
+    activation: 'softmax'
+  }));
+  return baseModel;
+}
